refactor: extract generic ModelCard from AsrModelCard
Task 9 of multi-diarization-models plan. Extracts the group-agnostic
model card UI into ModelCard.tsx with a parametrizable activeUsageLabel
prop. AsrModelCard becomes a thin alias that supplies the transcription-
specific label. Diarization will use ModelCard directly with its own
label in task 10.
